servidor/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
APScheduler==3.10.4
asgiref==3.7.2
asyncpg==0.28.0
bcrypt==4.0.1
certifi==2023.7.22
click==8.1.7
colorama==0.4.6
coverage==7.3.2
distlib==0.3.7
dnspython==2.4.2
EasyProcess==1.1
ecdsa==0.18.0
email-validator==2.0.0.post2
entrypoint2==1.1
fastapi==0.103.2
filelock==3.12.4
greenlet==3.0.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
keyboard==0.13.5
logging==0.4.9.6
MouseInfo==0.1.3
mss==9.0.1
numpy==1.26.0
opencv-python==4.8.1.78
packaging==23.2
passlib==1.7.4
Pillow==10.0.1
platformdirs==3.11.0
pluggy==1.3.0
ply==3.11
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.0
PyAutoGUI==0.9.54
pydantic==2.4.2
pydantic_core==2.10.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
pytest==7.4.2
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytweening==1.0.7
pytz==2023.3.post1
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
sqlparse==0.4.4
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
uvicorn==0.23.2
virtualenv==20.24.5
watchfiles==0.21.0
websockets==11.0.3
